// app/types.ts
export enum MemberCreateError {
	NONE = -1,
	INCORRRECT_CREDENTIALS = 0,
	SERVER_ERROR = 1,
	PASSWORD_EXPIRED = 2,
	INVALID_SESSION_ID = 3,
	UNKOWN_SERVER_ERROR = 4,
	DATABASE_ERROR = 5
}
// ...
export namespace Permissions {
	export enum FlightAssign {
		NO,
		YES
	}

	export enum MusterSheet {
		NO,
		YES
	}

	export enum PTSheet {
		NO,
		YES
	}

	export enum PromotionManagement {
		NONE,
		FULL
	}

	export enum AssignTasks {
		NO,
		YES
	}

	export enum AdministerPT {
		NO,
		YES
	}

	export enum ManageEvent {
		NONE,
		ADDDRAFTEVENTS,
		FULL
	}

	export enum EventContactSheet {
		NO,
		YES
	}

	export enum ORMOPORD {
		NO,
		YES
	}

	export enum AssignTemporaryDutyPosition {
		NO,
		YES
	}

	export enum ProspectiveMemberManagement {
		NONE,
		FULL
	}

	export enum EventLinkList {
		NO,
		YES
	}

	export enum ManageTeam {
		NONE,
		FULL
	}

	export enum FileManagement {
		NONE,
		FULL
	}

	export enum PermissionManagement {
		NONE,
		FULL
	}

	export enum DownloadCAPWATCH {
		NO,
		YES
	}

	export enum Notify {
		NO,
		GLOBAL
	}

	export enum RegistryEdit {
		NO,
		YES
	}

	export enum ScanAdd {
		NO,
		YES
	}
}
// ...
export interface NoSQLDocument {
	/**
	 * The ID assigned by NoSQL/MySQL
	 */
	_id?: string;
}
// ...
export interface Identifiable {
	/**
	 * The ID an object has. This is different from a NoSQL ID, as most of these
	 * are set by NHQ or by us
	 */
	id: string | number;
}

/**
 * Says that there is another ID to listen to, the account ID
 */
export interface AccountIdentifiable extends Identifiable {
	/**
	 * The ID of the account the folder belongs to
	 */
	accountID: string;
}
// ...
export interface RawAccountObject extends Identifiable, NoSQLDocument, NewAccountObject {
	/**
	 * The Account ID
	 */
	id: string;
	/**
	 * The ID of the main Google calendar
	 */
	mainCalendarID: string;
	/**
	 * The ID of the wing Google calendar
	 */
	wingCalendarID: string;
	/**
	 * The email address of the Google service account.  Not directly used by
	 * CAPUnit.com, however it is good to have for configuration.
	 */
	serviceAccount: MaybeObj<string>;
	/**
	 * The web link used to share the Google Calendar to Wing or others
	 * (only contains events with the publishToWingCalendar property true)
	 */
	shareLink: string;
	/**
	 * Initial password used to generate Google account for a unit
	 */
	initialPassword: string;
	/**
	 * Miscellaneous comments regarding the account
	 */
	comments: string;
	/**
	 * Whether or not the account is an echelon account
	 */
	echelon: boolean;
	/**
	 * The main organization of the account
	 */
	mainOrg: number;
	/**
	 * The ids of the organizations
	 */
	orgIDs: number[];
	/**
	 * Whether the account is a paid account
	 */
	paid: boolean;
	/**
	 * Datetime when the account expires in
	 */
	expires: number;
	/**
	 * How many events can be used if this account is paid for
	 */
	paidEventLimit: number;
	/**
	 * How many events can be used if this account is unpaid for
	 */
	unpaidEventLimit: number;
	/**
	 * Allow for website accounts to alias themselves
	 *
	 * Fake account IDs, as it were
	 *
	 * Will not be editable by the public
	 */
	aliases: string[];
}

/**
 * The object that splits all of the data used by different people to limit their views
 * to what they want and what they have permissions for
 */
export interface AccountObject extends RawAccountObject {
	/**
	 * Whether the account is a valid paid account
	 *
	 * Valid paid means it is paid for and not expired
	 */
	validPaid: boolean;
	/**
	 * Whether the account is expired
	 */
	expired: boolean;
}
// ...
export interface CAPMemberContact {
	/**
	 * A contact method to use to get in touch with the member
	 */
	ALPHAPAGER: CAPMemberContactInstance;
	/**
	 * A contact method to use to get in touch with the member
	 */
	ASSISTANT: CAPMemberContactInstance;
	/**
	 * A contact method to use to get in touch with the member
	 */
	CADETPARENTEMAIL: CAPMemberContactInstance;
	/**
	 * A contact method to use to get in touch with the member
	 */
	CADETPARENTPHONE: CAPMemberContactInstance;
	/**
	 * A contact method to use to get in touch with the member
	 */
	CELLPHONE: CAPMemberContactInstance;
	/**
	 * A contact method to use to get in touch with the member
	 */
	DIGITALPAGER: CAPMemberContactInstance;
	/**
	 * A contact method to use to get in touch with the member
	 */
	EMAIL: CAPMemberContactInstance;
	/**
	 * A contact method to use to get in touch with the member
	 */
	HOMEFAX: CAPMemberContactInstance;
	/**
	 * A contact method to use to get in touch with the member
	 */
	HOMEPHONE: CAPMemberContactInstance;
	/**
	 * A contact method to use to get in touch with the member
	 */
	INSTANTMESSAGER: CAPMemberContactInstance;
	/**
	 * A contact method to use to get in touch with the member
	 */
	ISDN: CAPMemberContactInstance;
	/**
	 * A contact method to use to get in touch with the member
	 */
	RADIO: CAPMemberContactInstance;
	/**
	 * A contact method to use to get in touch with the member
	 */
	TELEX: CAPMemberContactInstance;
	/**
	 * A contact method to use to get in touch with the member
	 */
	WORKFAX: CAPMemberContactInstance;
	/**
	 * A contact method to use to get in touch with the member
	 */
	WORKPHONE: CAPMemberContactInstance;
}

/**
 * Used for looping through the contact types
 */
export type CAPMemberContactType = keyof CAPMemberContact;

/**
 * Used for getting priorities and such
 */
export type CAPMemberContactPriority = keyof CAPMemberContactInstance;

/**
 * The permissions list for various members
 */
export interface MemberPermissions {
	// Start the Cadet Staff permissions
	/**
	 * Whether or not the user can assign flight members
	 */
	FlightAssign: Permissions.FlightAssign;
	/**
	 * Whether or not the user can get the muster sheet
	 */
	MusterSheet: Permissions.MusterSheet;
	/**
	 * Whether or not the user can get PT sheets
	 */
	PTSheet: Permissions.PTSheet;
	/**
	 * Whether or not the user can manage promotions
	 */
	PromotionManagement: Permissions.PromotionManagement;
	/**
	 * Whether or not the user can assign tasks
	 */
	AssignTasks: Permissions.AssignTasks;
	/**
	 * Whether or not the user can administer PT
	 */
	AdministerPT: Permissions.AdministerPT;

	// Start Manager permissions
	/**
	 * Whether or not the user can manage events and to what degree
	 */
	ManageEvent: Permissions.ManageEvent;
	/**
	 * Whether or not the user can get event contact information
	 */
	EventContactSheet: Permissions.EventContactSheet;
	/**
	 * Whether or not the user can get ORM OPORD information
	 */
	ORMOPORD: Permissions.ORMOPORD;
	/**
	 * Whether or not the user can assign temporary duty positions
	 */
	AssignTemporaryDutyPositions: Permissions.AssignTemporaryDutyPosition;

	// Admin privileges
	/**
	 * Whether or not the user can manage prospective members
	 */
	ProspectiveMemberManagement: Permissions.ProspectiveMemberManagement;
	/**
	 * Whether or not the user can view a list of all events
	 */
	EventLinkList: Permissions.EventLinkList;
	/**
	 * Whether or not the user can add a team
	 */
	ManageTeam: Permissions.ManageTeam;
	/**
	 * Whether or not the user can manage files
	 */
	FileManagement: Permissions.FileManagement;
	/**
	 * Whether or not the user can manage permissions of others
	 */
	PermissionManagement: Permissions.PermissionManagement;
	/**
	 * Whether or not the user can download CAPWATCH files
	 */
	DownloadCAPWATCH: Permissions.DownloadCAPWATCH;
	/**
	 * Whether or not the member can create banner notifications
	 */
	CreateNotifications: Permissions.Notify;
	/**
	 * Whether or not members can initiate scan add sessions
	 */
	ScanAdd: Permissions.ScanAdd;

	// Developer/super admin privileges?
	// Will change when utilities are more user friendly
	/**
	 * Whether or not the user can edit the registry
	 */
	RegistryEdit: Permissions.RegistryEdit;
}

export interface StoredMemberPermissions extends NoSQLDocument {
	member: MemberReference;
	accountID: string;
	permissions: MemberPermissions;
}

/**
 * This information is used to store if a member has said they will be absent until a date,
 * and comments associated with it
 */
export interface AbsenteeInformation {
	/**
	 * Used by the member to describe why a cadet will be absent
	 */
	comments: string;
	/**
	 * Used to say how long the cadet is absent for
	 */
	absentUntil: number;
}

/**
 * String representation of permissions
 */
export type MemberPermission = keyof MemberPermissions;

/**
 * String representation of the member type, as there needs to be a way to differentiate
 */
export type MemberType = CAPMemberType;

/**
 * Information stored about a member in our database
 *
 * Does not include permission information, which is stored seperately
 * This is so that permissions can be on an account basis
 */
export interface RawMemberObject extends Identifiable {
	/**
	 * The CAPID of the member
	 */
	id: number | string;
	/**
	 * Contact information for the user
	 */
	contact: CAPMemberContact;
	/**
	 * First name of member
	 */
	nameFirst: string;
	/**
	 * Middle name of member
	 */
	nameMiddle: string;
	/**
	 * Last name of member
	 */
	nameLast: string;
	/**
	 * Suffix of member
	 */
	nameSuffix: string;
	/**
	 * User login ID
	 */
	usrID: string;
	/**
	 * The type of user, as there are multiple
	 */
	type: MemberType;
	/**
	 * Used to easily reference teams
	 */
	teamIDs: number[];
	/**
	 * Shows how long the member is absent for
	 *
	 * Should not be used if null or if the time has passed
	 */
	absenteeInformation: AbsenteeInformation | null;
}

/**
 * Describes a member
 *
 * CAPWATCHMember and NHQMember have their sources in NHQ
 * ProspectiveMembers are located in our database
 */
export interface MemberObject extends RawMemberObject {
	/**
	 * Tells what the member can do
	 */
	permissions: MemberPermissions;
}
// ...
export interface ProspectiveMemberObject extends RawProspectiveMemberObject, CAPMemberObject {
	/**
	 * Prospective members have string IDs
	 */
	id: string;
	/**
	 * Typescript deliminator
	 */
	type: 'CAPProspectiveMember';
	/**
	 * Shows what the member is allowed to do
	 */
	permissions: MemberPermissions;
}
// ...
export type Member = ProspectiveMemberObject | NHQMemberObject;
// ...
export interface SuccessfulSigninReturn {
	/**
	 * No error with this session
	 */
	error: MemberCreateError.NONE;
	/**
	 * The member details
	 */
	member: Member;
	/**
	 * The ID for the session
	 */
	sessionID: string;
	/**
	 * Returns the amount of notifications the member has
	 */
	notificationCount: number;
	/**
	 * Returns the amount of unfinished tasks the member has
	 */
	taskCount: number;
}

/**
 * Used to create a new password
 */
export interface ExpiredSuccessfulSigninReturn {
	/**
	 * Just an expired password
	 */
	error: MemberCreateError.PASSWORD_EXPIRED;
	/**
	 * A session ID to set a new password
	 */
	sessionID: string;
}

/**
 * Can't simply return null, need to return information as to why
 * things failed
 */
export interface FailedSigninReturn {
	/**
	 * Contains error details
	 */
	error:
		| MemberCreateError.INCORRRECT_CREDENTIALS
		| MemberCreateError.SERVER_ERROR
		| MemberCreateError.UNKOWN_SERVER_ERROR
		| MemberCreateError.INVALID_SESSION_ID;
}

/**
 * Allows for multiplexing the data together but still have type inference and
 * not use try/catch
 */
export type SigninReturn =
	| SuccessfulSigninReturn
	| ExpiredSuccessfulSigninReturn
	| FailedSigninReturn;
// ...
export type Timezone =
	| 'America/New_York'
	| 'America/Chicago'
	| 'America/Denver'
	| 'America/Los_Angeles'
	| 'America/Arizona'
	| 'America/Anchorage'
	| 'America/Hawaii'
	| 'America/Puerto_Rico';

export interface WebsiteInformation {
	/**
	 * What the website is called
	 */
	Name: string;
	/**
	 * -, ::, etc. Personal taste, used in the title of the page
	 */
	Separator: string;
	/**
	 * Controls how many events will show up in the 'upcoming' page
	 */
	ShowUpcomingEventCount: number;
	/**
	 * How many images are downloaded when scrolling in the photo library
	 */
	PhotoLibraryImagesPerPage: number;
	/**
	 * What timezone the unit is operating within
	 */
	Timezone: Timezone;
}

/**
 * The website contact information the webmaster may want to set
 *
 * All of this information is currently only used on the footer of
 * the website
 */
export interface WebsiteContact {
	/**
	 * A Facebook handle that may want to be used
	 */
	FaceBook: null | string;
	/**
	 * A Twitter handle
	 */
	Twitter: null | string;
	/**
	 * A link to a YouTube account a squadron may have
	 */
	YouTube: null | string;
	/**
	 * A link to a LinkedIn account that can be used
	 */
	LinkedIn: null | string;
	/**
	 * ditto Instagram
	 */
	Instagram: null | string;
	/**
	 * ditto Flickr
	 */
	Flickr: null | string;
	/**
	 * Invite link for a Discord server
	 */
	Discord: null | string;
	/**
	 * This is the place where normal meetings take place
	 *
	 * Only used for style at the bottom of the page, nothing else actually uses this
	 *
	 * Maybe could be the default place for Events?
	 */
	MeetingAddress: null | {
		Name: string;
		FirstLine: string;
		SecondLine: string;
	};
	/**
	 * The place to mail stuff to, may not necessarily match the meeting address
	 */
	MailingAddress: null | {
		Name: string;
		FirstLine: string;
		SecondLine: string;
	};
}

/**
 * Used to configure the different flights
 */
export interface RankAndFileInformation {
	/**
	 * Describes the flights that an account may field, e.g. Alpha Bravo Charlie Delta
	 */
	Flights: string[];
}

/**
 * Each account has a registry, stores configuration
 */
export interface RegistryValues extends NoSQLDocument, AccountIdentifiable {
	/**
	 * How to contact the account; email, social media, etc;
	 */
	Contact: WebsiteContact;
	/**
	 * Website naming details
	 */
	Website: WebsiteInformation;
	/**
	 * Contains information about the different flights
	 */
	RankAndFile: RankAndFileInformation;
}